# Optimize canvas hit detection by rendering features in a limited extent