@tschaub 2015-01-08
@elemoine did you want to give this another look?  @tsauerwein and I agree it is good to go.