@elemoine 2015-01-08
Where are the tests? ;-) Just kidding. I don't really think this needs tests at this point. Please merge.